MW satellite LF
A version of SAM, which is calibrated to the field LF at z=0, is
applied to 31 MW halo merger trees. The figure shows the predicted
r-band LF of satellite galaxies and compares the predictions with data.
Each grey line is a LF of a simulated MW halo. The black is the average
LF of all the simulated MW halos. The black error bars are the expected
error of the mean, assuming it is poisson. The red band shows the
standard deviation among the simulated halos. The red error bars show
the expected Poisson error.  The grey lines with dark green dashed
lines overlaid are the LFs of the simulated  systems where at least 2
satellites are brighter than M_r=-16.5. The light green is the average
of those. The blue line with error bar is the observed MW LF.