corrected version of MWSatMAH.pdf
I show the mass accretion histories of halos and sub halos in 38 MW
host simulations. The red line in each figure is the MAH of the MW host
halo. The black lines are the MAHs of sub halos in each host that are
more massive than 10^9Msun at z=0. The orange lines in some of the
figures are the MAH of the sub halos that host a satellite galaxies
more luminous than SMC (Mr<-16.5). There are 8 simulated hosts having 2
(or more) satellite galaxies as LMC/SMC analogs as predicted by my SAM.
The figures of those hosts are indicated by “MW/LMC+SMC”. In 5 (out of
8) of them, only one of the LMC/SMC satellites has its host sub host
still existing in the halo catalog. The other LMC/SMC is an ‘orphan’
galaxy with no existing sub halo. For each MAH, the dashed portion of
the line indicate the period when the sub halo is located within the
viral radius of the final host.